chore(napi): mark tsfn data as pub and split SendableResolver to indent file
