+++ crates/napi/src/tokio_runtime.rs
@@ -6,7 +6,7 @@ use tokio::runtime::Runtime;
 
 use crate::{bindgen_runtime::ToNapiValue, sys, Env, Error, Result};
 #[cfg(not(feature = "noop"))]
-use crate::{JsDeferred, Unknown};
+use crate::{JsDeferred, SendableResolver, Unknown};
 
 #[cfg(not(feature = "noop"))]
 fn create_runtime() -> Runtime {
@@ -168,41 +168,6 @@ pub fn within_runtime_if_available<F: FnOnce() -> T, T>(f: F) -> T {
 #[cfg(feature = "noop")]
 pub fn within_runtime_if_available<F: FnOnce() -> T, T>(f: F) -> T {
   f()
-}
-
-#[cfg(not(feature = "noop"))]
-pub struct SendableResolver<
-  Data: 'static + Send,
-  R: 'static + FnOnce(sys::napi_env, Data) -> Result<sys::napi_value>,
-> {
-  inner: R,
-  _data: PhantomData<Data>,
-}
-
-#[cfg(not(feature = "noop"))]
-// the `SendableResolver` will be only called in the `threadsafe_function_call_js` callback
-// which means it will be always called in the Node.js JavaScript thread
-// so the inner function is not required to be `Send`
-// but the `Send` bound is required by the `execute_tokio_future` function
-unsafe impl<Data: 'static + Send, R: 'static + FnOnce(sys::napi_env, Data) -> Result<sys::napi_value>>
-  Send for SendableResolver<Data, R>
-{
-}
-
-#[cfg(not(feature = "noop"))]
-impl<Data: 'static + Send, R: 'static + FnOnce(sys::napi_env, Data) -> Result<sys::napi_value>>
-  SendableResolver<Data, R>
-{
-  pub fn new(inner: R) -> Self {
-    Self {
-      inner,
-      _data: PhantomData,
-    }
-  }
-
-  pub fn resolve(self, env: sys::napi_env, data: Data) -> Result<sys::napi_value> {
-    (self.inner)(env, data)
-  }
 }
 
 #[cfg(feature = "noop")]
